backend/worker/worker_requirements.txt
SQLAlchemy==1.3.20
psycopg2-binary==2.8.6
python-crontab==2.5.1
rq==1.9.0
rq-dashboard==0.3.12
click==7.1.2
requests==2.22.0
configparser==3.5.3
openpyxl==3.0.5
XlsxWriter==1.4.3
cryptography==2.8
asyncssh==2.4.2
cx-Oracle==8.2.1
pandas==1.1.5
